Rename ClientStateAddressResolver to KeyStateAddressResolver
